Backend/singularity_requirements.txt
absl-py==1.3.0
alembic==1.8.1
antlr4-python3-runtime==4.8
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
certipy==0.1.3
conda==22.9.0
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
einops==0.7.0
entrypoints==0.4
executing==1.1.1
fairseq==0.12.2
fastjsonschema==2.16.2
fonttools==4.38.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
greenlet==1.1.3.post0
grpcio==1.50.0
hydra-core==1.0.7
imageio==2.22.2
importlib-metadata==5.0.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter-telemetry==0.1.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterhub==3.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.6.0.66
packaging==21.3
pamela==1.0.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacrebleu==2.3.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.5.1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
torch==1.13.0+cu116
torch-tb-profiler==0.4.0
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
youtube-dl==2021.12.17
zipp==3.10.0